+++ src/config.lua
@@ -128,6 +128,9 @@ local config = {
 	EphyraOverhaul = {
 		Enabled = true
 	},
+	ExtraLastStandsFirst = {
+		Enabled = true
+	},
 }
 
 local description = {
@@ -260,6 +263,9 @@ local description = {
 	},
 	EphyraOverhaul = {
 		Enabled = "Enable to overhaul Ephyra. 6 Rooms containing pylons are designated, and shown with PonyQOL2 installed. Completing rooms with no pylon does not increase the pylon counter, letting you complete more than 6. Every door in Ephyra is open and contains a reward. Completing the 6 pylon rooms unseals the boss room and locks all doors as normal. If you complete more than 6 rooms, each room you enter beyond the 6th will contain reinforcements from Tartarus."
+	},
+	ExtraLastStandsFirst = {
+		Enabled = "Enable to make extra last stands be used first."
 	},
 }