+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added a band for heroic rarity keepsakes
+- Added ExtraLastStandsFirst
+
+### Fixed
+
+- Fixed issues with the Artificer card change
+
 ## [0.4.2] - 2024-06-08
 
 ### Fixed